requirements.txt
diffusers
fastapi
Pillow
pydantic
starlette
torch
uvicorn[standard]
transformers
accelerate
xformers